backend/requirements.txt
Flask==2.0.1
Pillow==10.0.0